New patch 2 fix a bug in codec dynamic loading & linking
If one or several codecs (e.g. libflac) are not installed in the system
and no logging (-d key) specified in command line, After a load of one of the codecs failed (file missing),
the rest of codecs and liblirc_client do not load too, despite the fact that the libraries are present.

If right after successful call of dlopen() function make dummy call of dlerror() function, the problem disappears